--- a/atoms/Cargo.toml
+++ b/atoms/Cargo.toml
@@ -10,7 +10,7 @@ description = "Atoms for the swc project."
 edition = "2018"
 
 [dependencies]
-string_cache = "0.7"
+string_cache = "0.8"
 
 [build-dependencies]
-string_cache_codegen = "0.4"
+string_cache_codegen = "0.5"