--- a/common/Cargo.toml
+++ b/common/Cargo.toml
@@ -16,7 +16,7 @@ fold = ["ast_node/fold"]
 
 [dependencies]
 ast_node = { version = "0.4", path = "../macros/ast_node" }
-string_cache = "0.7"
+string_cache = "0.8"
 chashmap = "2.2"
 either = "1.5"
 scoped-tls = { version = "1" }