--- a/ecmascript/parser/Cargo.toml
+++ b/ecmascript/parser/Cargo.toml
@@ -21,7 +21,7 @@ swc_common = { version = "0.3.2", path ="../../common" }
 swc_ecma_ast = { version = "0.9.3", path ="../ast" }
 swc_ecma_parser_macros = { version = "0.3", path ="./macros" }
 enum_kind = { version = "0.1", path ="../../macros/enum_kind" }
-unicode-xid = "0.1"
+unicode-xid = "0.2"
 log = { version = "0.4", features = ["release_max_level_debug"] }
 either = { version = "1.4" }
 serde = { version = "1", features = ["derive"] }