--- a/ecmascript/parser/src/parser/stmt/mod.rs
+++ b/ecmascript/parser/src/parser/stmt/mod.rs
@@ -281,7 +281,7 @@ impl<'a, I: Tokens> Parser<'a, I> {
 
         if self.ctx().strict {
             match *expr {
-                Expr::Ident(Ident { ref sym, span, .. }) => match sym {
+                Expr::Ident(Ident { ref sym, span, .. }) => match *sym {
                     js_word!("enum") | js_word!("interface") => {
                         self.emit_err(span, SyntaxError::InvalidIdentInStrict);
                     }