--- a/ecmascript/transforms/src/compat/es2015/classes/native.rs
+++ b/ecmascript/transforms/src/compat/es2015/classes/native.rs
@@ -7,7 +7,7 @@ macro_rules! native {
             $i:tt
         ),*
     ) => {
-        match $sym{
+        match *$sym{
             $(
                 js_word!($i) => true,
             )*