--- a/ecmascript/transforms/src/modules/common_js/tests.rs
+++ b/ecmascript/transforms/src/modules/common_js/tests.rs
@@ -536,7 +536,7 @@ export {foo as default, bar};
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.default = exports.bar = void 0;
+exports.bar = exports.default = void 0;
 var foo, bar;
 exports.default = foo;
 exports.bar = bar;