--- a/ecmascript/transforms/src/modules/umd/tests.rs
+++ b/ecmascript/transforms/src/modules/umd/tests.rs
@@ -562,7 +562,7 @@ export {foo as default, bar};
   Object.defineProperty(_exports, "__esModule", {
     value: true
   });
-  _exports.default = _exports.bar = void 0;
+  _exports.bar = _exports.default = void 0;
   var foo, bar;
   _exports.default = foo;
   _exports.bar = bar;