```
test angular          ... bench:  30,197,806 ns/iter (+/- 1,644,729) = 23 MB/s
test angular_ts       ... bench:  32,007,512 ns/iter (+/- 519,524) = 22 MB/s
test backbone         ... bench:   4,270,062 ns/iter (+/- 101,528) = 14 MB/s
test backbone_ts      ... bench:   4,540,125 ns/iter (+/- 129,196) = 13 MB/s
test colors           ... bench:      52,109 ns/iter (+/- 9,868) = 22 MB/s
test colors_ts        ... bench:      54,065 ns/iter (+/- 9,597) = 21 MB/s
test jquery           ... bench:  24,359,390 ns/iter (+/- 3,973,410) = 11 MB/s
test jquery_ts        ... bench:  25,600,275 ns/iter (+/- 361,245) = 10 MB/s
test jquery_mobile    ... bench:  39,791,297 ns/iter (+/- 16,603,058) = 11 MB/s
test jquery_mobile_ts ... bench:  39,795,432 ns/iter (+/- 1,418,453) = 11 MB/s
test mootools         ... bench:  19,675,621 ns/iter (+/- 293,130) = 8 MB/s
test mootools_ts      ... bench:  20,628,544 ns/iter (+/- 1,725,182) = 7 MB/s
test underscore       ... bench:   3,631,278 ns/iter (+/- 129,901) = 11 MB/s
test underscore_ts    ... bench:   3,997,078 ns/iter (+/- 107,128) = 10 MB/s
test yui              ... bench:  18,224,793 ns/iter (+/- 4,342,498) = 18 MB/s
test yui_ts           ... bench:  19,406,389 ns/iter (+/- 366,219) = 17 MB/s
```